# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/auth/logout": {
            "get": {
                "description": "登出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "登出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/upload": {
            "post": {
                "description": "批量上传图片",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "批量上传图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "图片",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category",
                        "in": "formData"
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/public/images": {
            "post": {
                "description": "分页列出图片",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "分页列出图片",
                "parameters": [
                    {
                        "description": "分页",
                        "name": "page",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "图片列表",
                        "schema": {
                            "$ref": "#/definitions/common.PageResp"
                        }
                    }
                }
            }
        },
        "/api/public/images/{name}": {
            "get": {
                "description": "根据文件名获取图片",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "根据文件名获取图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件名",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "图片内容",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/public/login": {
            "post": {
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handles.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "$ref": "#/definitions/handles.LoginResp"
                        }
                    }
                }
            }
        },
        "/api/public/random": {
            "get": {
                "description": "随机获取一个图片 支持分类",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "随机获取一个图片 支持分类",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "图片内容",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/public/short/{short_link}": {
            "get": {
                "description": "根据短链获取图片",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "根据短链获取图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链",
                        "name": "short_link",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "图片内容",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.PageResp": {
            "type": "object",
            "properties": {
                "content": {},
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "handles.LoginReq": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handles.LoginResp": {
            "type": "object",
            "properties": {
                "expire": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "model.PageReq": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                }
            }
        },
...
    }
}